kiratech/requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
pytz==2023.3
sniffio==1.3.0
sqlparse==0.4.4
urllib3==2.0.2
